pyproject.toml
[tool.poetry]
name = "telegram-notifier"
version = "0.1.0"
description = ""
authors = ["laraib_sidd <laraib.sidd7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
python-telegram-bot = "13.1"
pytz = "^2023.3.post1"
pymongo = "^4.6.0"
praw = "^7.7.1"
xmltodict = "^0.13.0"
beautifulsoup4 = "^4.12.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

